manifest: move filterSupportedRepresentation in src/manifest/adaptation.ts
